Struct bsn1::Der

pub struct Der { /* fields omitted */ }

Der owns DerRef and represents DER.

Implementations

impl Der

pub fn new(id: &IdRef, contents: &[u8]) -> Self

Creates a new instance from id and contents .

Warnings

ASN.1 does not allow some universal identifier for DER, however, this function accepts such an identifier. For example, ‘Octet String’ must be primitive in DER, but this function will construct a new instance even if id represenets constructed ‘Octet String.’

Examples

use bsn1::{Der, IdRef};

let id = IdRef::octet_string();
let contents: &[u8] = &[10, 20, 30];

let der = Der::new(id, contents);

assert_eq!(id, der.id());
assert_eq!(contents, der.contents());

pub fn from_bytes(bytes: &[u8]) -> Result<Self, Error>

Parses bytes starting with DER octets and builds a new instance.

This function ignores extra octet(s) at the end of bytes if any.

This function is same to TryFrom::try_from .

Warnings

ASN.1 does not allow some universal identifier for DER, however, this function accepts such an identifier. For example, ‘Octet String’ must be primitive in DER, but this function returns Ok for constructed Octet String DER.

pub unsafe fn from_bytes_unchecked(bytes: &[u8]) -> Self

Builds a new instance holding bytes without any sanitization.

bytes must not include any extra octet.

If it is not sure whether bytes are valid octets as an ‘DER’ or not, use TryFrom implementation or from_bytes.

The difference from from_bytes_starts_with_unchecked is that from_bytes_starts_with_unchecked checks the ‘LENGTH’ octets and excludes extra octet(s) at the end if any while this method does not check at all (i.e. from_bytes_starts_with_unchecked allows extra octets at the end.)

Safety

The behavior is undefined if bytes is not formatted as a DER.

Examples

use bsn1::{Der, IdRef};

let der0 = Der::new(IdRef::octet_string(), &[]);
let der1 = unsafe { Der::from_bytes_unchecked(der0.as_ref()) };
assert_eq!(der0, der1);

pub unsafe fn from_bytes_starts_with_unchecked(bytes: &[u8]) -> Self

Builds a new instance from bytes that starts with ‘DER’ octets.

bytes may include some extra octet(s) at the end.

If it is not sure whether bytes starts with DER octets or not, use TryFrom implementation or from_bytes .

The difference from from_bytes_unchecked is that this function checks the ‘LENGTH’ octets and excludes extra octet(s) at the end if any, while from_bytes_unchecked does not check at all.

Safety

The behavior is undefined if bytes does not start with ‘ASN.1 DER’ octets.

Examples

use bsn1::{Der, IdRef};

let der0 = Der::new(IdRef::octet_string(), &[]);
let mut bytes = Vec::from(der0.as_ref() as &[u8]);
bytes.extend(&[1, 2, 3]);

let der1 = unsafe { Der::from_bytes_starts_with_unchecked(bytes.as_ref()) };
assert_eq!(der0, der1);

pub fn from_id_iterator<I>(id: &IdRef, contents: I) -> Self where
I: Iterator + Clone,
I::Item: AsRef<[u8]>,

Creates a new instance from id and contents .

Warnings

ASN.1 does not allow some universal identifier for DER, however, this function accepts such an identifier. For example, ‘Octet String’ must be primitive in DER, but this function will construct a new instance even if id represenets constructed ‘Octet String.’

Examples

use bsn1::{Der, IdRef};

let id = IdRef::sequence();

// Build instance using function 'from_id_iterator()'.
let contents: &[Der] = &[Der::utf8_string("foo"), Der::integer(29)];
let der = Der::from_id_iterator(id, contents.iter());

// Build instance using function 'new()'.
let contents: Vec<u8> = contents.iter()
                        .map(|i| Vec::from(i.as_ref() as &[u8]))
                        .flatten().collect();
let expected = Der::new(id, &contents);

assert_eq!(expected, der);

pub fn boolean(val: bool) -> Self

Returns a new instance representing boolean.

Examples

use bsn1::{contents, Der, IdRef};

let val = true;
let der = Der::boolean(val);

assert_eq!(IdRef::boolean(), der.id());
assert_eq!(val, contents::to_bool_der(der.contents()).unwrap());

pub fn integer<T>(val: T) -> Self where
T: PrimInt,

Returns a new instance representing integer.

Type T should be the builtin primitive integer types (e.g., u8, u32, isize, i128, …)

Examples

use bsn1::{contents, Der, IdRef};

let val = 39;
let der = Der::integer(val);

assert_eq!(IdRef::integer(), der.id());
assert_eq!(val, contents::to_integer(der.contents()).unwrap());

pub fn utf8_string(val: &str) -> Self

Returns a new instance representing utf8_string.

Examples

use bsn1::{Der, IdRef};

let val = &"foo";
let der = Der::utf8_string(val);

assert_eq!(IdRef::utf8_string(), der.id());
assert_eq!(val.as_bytes(), der.contents());

pub fn octet_string(val: &[u8]) -> Self

Returns a new instance representing octet_string.

Examples

use bsn1::{Der, IdRef};

let val = &[1, 2, 3];
let der = Der::octet_string(val);

assert_eq!(IdRef::octet_string(), der.id());
assert_eq!(val, der.contents());

impl Der

pub fn into_vec(self) -> Vec<u8>

Consumes self , returning Vec .

Examples

use bsn1::{Der, IdRef};

let id = IdRef::octet_string();
let contents: &[u8] = &[0, 1, 2, 3, 4];

let der = Der::new(id, contents);
let v = der.clone().into_vec();

assert_eq!(der.as_ref() as &[u8], v.as_ref() as &[u8]);